pyproject.toml
[project]
name = "aiogrambot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "aiosqlite>=0.22.1",
    "lxml>=6.0.2",
    "lxml-stubs>=0.5.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "html2image==2.0.7",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
]
